+++ src/compas_rhino/artists/boxartist.py
@@ -18,10 +18,43 @@ class BoxArtist(ShapeArtist):
     -----
     See :class:`compas_rhino.artists.ShapeArtist` for all other parameters.
 
+    Examples
+    --------
+    .. code-block:: python
+
+        import random
+        from compas.geometry import Pointcloud
+        from compas.geometry import Box
+        from compas.utilities import i_to_rgb
+
+        import compas_rhino
+        from compas_rhino.artists import BoxArtist
+
+        pcl = Pointcloud.from_bounds(10, 10, 10, 100)
+        tpl = Box.from_width_height_depth(0.3, 0.3, 0.3)
+
+        compas_rhino.clear_layer("Test::BoxArtist")
+
+        for point in pcl.points:
+            box = tpl.copy()
+            box.frame.point = point
+            artist = BoxArtist(box, color=i_to_rgb(random.random()), layer="Test::BoxArtist")
+            artist.draw()
     """
 
-    def draw(self):
+    def draw(self, show_vertices=False, show_edges=False, show_faces=True, join_faces=True):
         """Draw the box associated with the artist.
+
+        Parameters
+        ----------
+        show_vertices : bool, optional
+            Default is ``False``.
+        show_edges : bool, optional
+            Default is ``False``.
+        show_faces : bool, optional
+            Default is ``True``.
+        join_faces : bool, optional
+            Default is ``True``.
 
         Returns
         -------
@@ -29,14 +62,22 @@ class BoxArtist(ShapeArtist):
             The GUIDs of the objects created in Rhino.
         """
         vertices = [list(vertex) for vertex in self.shape.vertices]
-        faces = self.shape.faces
-        edges = self.shape.edges
-        points = [{'pos': point, 'color': self.color} for point in vertices]
-        lines = [{'start': vertices[i], 'end': vertices[j], 'color': self.color} for i, j in edges]
-        polygons = [{'points': [vertices[index] for index in face], 'color': self.color} for face in faces]
-        guids = compas_rhino.draw_points(points, layer=self.layer, clear=False, redraw=False)
-        guids += compas_rhino.draw_lines(lines, layer=self.layer, clear=False, redraw=False)
-        guids += compas_rhino.draw_faces(polygons, layer=self.layer, clear=False, redraw=False)
+        guids = []
+        if show_vertices:
+            points = [{'pos': point, 'color': self.color, 'name': self.name} for point in vertices]
+            guids += compas_rhino.draw_points(points, layer=self.layer, clear=False, redraw=False)
+        if show_edges:
+            edges = self.shape.edges
+            lines = [{'start': vertices[i], 'end': vertices[j], 'color': self.color, 'name': self.name} for i, j in edges]
+            guids += compas_rhino.draw_lines(lines, layer=self.layer, clear=False, redraw=False)
+        if show_faces:
+            faces = self.shape.faces
+            if join_faces:
+                guid = compas_rhino.draw_mesh(vertices, faces, layer=self.layer, name=self.name, color=self.color, disjoint=True)
+                guids.append(guid)
+            else:
+                polygons = [{'points': [vertices[index] for index in face], 'color': self.color, 'name': self.name} for face in faces]
+                guids += compas_rhino.draw_faces(polygons, layer=self.layer, clear=False, redraw=False)
         self._guids = guids
         return guids